Upload code used for developing the Shiny App
The whole bulk of the code developed in 2020 is uploaded in Github for
control version purposes.

As of this version, it contains an error with the xlsx library and it is
not running properly. It needs a fix in the local system regarding the
Java version used. If the Java installation fixes the issue, there will
not be a new commit. If a new excel-based package is required, then a
commit will follow to address the issue.